refactor(agent): split ui/agent-sidebar into ui/agent/{asks,messages,sidebar,greeting}
The original ui/agent-sidebar/ folder accumulated four distinct concerns
under one namespace: the sidebar itself (Header, Input, tabs, history),
the message list, the multi-question Asks block, and the hero greeting.
Splitting them into siblings under ui/agent/ matches the actual feature
boundaries and makes the kit easier to navigate.

File moves (git tracks renames):
- ui/agent-sidebar/AgentAsks.stories.tsx           → ui/agent/asks/
- ui/agent-sidebar/AgentSidebarMultiQuestion.tsx   → ui/agent/asks/
- ui/agent-sidebar/AgentMessages.stories.tsx       → ui/agent/messages/
- ui/agent-sidebar/AgentSidebarMessage.tsx         → ui/agent/messages/
- ui/agent-sidebar/AgentSidebarMessages.tsx        → ui/agent/messages/
- ui/agent-sidebar/AgentSidebar.{tsx,stories,test} → ui/agent/sidebar/
- ui/agent-sidebar/AgentSidebarHeader.tsx          → ui/agent/sidebar/
- ui/agent-sidebar/AgentSidebarInput.tsx           → ui/agent/sidebar/
- ui/agent-sidebar/{types,mock-data}.ts            → ui/agent/sidebar/
- ui/agent-sidebar/AgentGreeting.{tsx,stories,test} → ui/agent/greeting/

Import updates: cross-folder paths in the AgentSidebar barrel, sidebar
stories, mock-data, AgentSidebarMessages, plus the two external
importers (src/index.ts and AppShell.tsx). All public export names
unchanged — only path strings move.

Added meta export on AgentSidebarMultiQuestion.tsx as "AgentAsks" so
pnpm components validate continues to pass (the kit's components script
expects one meta-bearing file per component folder; this is now the
sole non-story file in asks/).

Greeting tuning in the same pass (visual feedback):
- Logo↔heading gap-4 → gap-2 (tighter pairing)
- IconButtons sm → md (hero context, more presence)
- Model picker pill h-8 → h-10 to align with the larger icon row

Verification:
- pnpm typecheck: clean
- pnpm components validate: All 52 component(s) have valid metadata
- 22 AgentGreeting tests + 7 AgentSidebar tests all pass at new paths
- 4 pre-existing ThemeProvider failures (happy-dom localStorage) unchanged
