+++ src/components/ui/agent/messages/AgentSidebarMessage.tsx
