+++ src/components/ui/agent/sidebar/AgentSidebar.test.tsx
