+++ src/components/ui/agent/sidebar/AgentSidebarHeader.tsx
